dynamic-mqtt-mapper/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
paho-mqtt==1.6.1
aiofiles==23.2.1